+++ projects/ontimize-web-ngx/src/lib/interfaces/auth-service.interface.ts
@@ -4,5 +4,4 @@ export interface IAuthService {
   startsession(user: string, password: string): Observable<string | number>;
   endsession(user: string, sessionId: string | number): Observable<number>;
   hassession(user: string, sessionId: string | number): Observable<boolean>;
-  redirectLogin?(sessionExpired?: boolean);
 }