+++ projects/ontimize-web-ngx/src/lib/interfaces/index.ts
@@ -9,7 +9,6 @@ export * from './form-control-component.interface';
 export * from './form-data-component.interface';
 export * from './form-data-type-component.interface';
 export * from './local-storage-component.interface';
-export * from './login-service.interface';
 export * from './multiple-selection.interface';
 export * from './o-context-menu.interface';
 export * from './o-form-layout-tab-group.interface';