+++ projects/ontimize-web-ngx/src/lib/pipes/index.ts
@@ -6,3 +6,4 @@ export * from './o-real.pipe';
 export * from './o-moment.pipe';
 export * from './o-currency.pipe';
 export * from './o-percentage.pipe';
+export * from './o-icon.pipe';